🧑‍💻🚚 Move util functions to seperate file
